Moved stack-frame into multiple files
